Registry is loaded from config dir
